requirements.txt
python-dotenv==0.6.0
win-unicode-console==0.5
vk-requests==0.9.6
